Added initial box parse logic (#37)
* Updated gitignore

* Working on intial parsing

* Added header parsing

* Working on file parsing

* Working on Parsing

* Added basic parsing

* Refactored to return status code

* Added function arg

+++ src/boxes/box.h
@@ -8,7 +8,7 @@
 
 class Box {
 public:
-    int size;
+    uint64_t size;
     std::string name;
 
     Box() {}
